+++ mqtt/src/com/splunk/modinput/mqtt/MQTTModularInput.java
@@ -105,6 +105,8 @@ public class MQTTModularInput extends ModularInput {
 
 		String messageHandlerImpl = DEFAULT_MESSAGE_HANDLER;
 		String messageHandlerParams = "";
+		
+		String activationKey = "";
 
 		Transport transport = getTransportInstance(params, stanzaName);
 
@@ -113,8 +115,11 @@ public class MQTTModularInput extends ModularInput {
 			if (value == null) {
 				continue;
 			}
-
-			if (param.getName().equals("topic_name")) {
+			
+			if (param.getName().equals("activation_key")) {
+				activationKey = param.getValue();
+			} 
+			else if (param.getName().equals("topic_name")) {
 				topicName = param.getValue();
 			} else if (param.getName().equals("broker_host")) {
 				brokerHost = param.getValue();
@@ -179,6 +184,10 @@ public class MQTTModularInput extends ModularInput {
 				setJVMSystemProperties(param.getValue());
 			}
 
+		}
+		
+		if(!activationKeyCheck(activationKey,"MQTT Modular Input")){
+			System.exit(2);
 		}
 
 		String brokerURL = brokerProtocol + brokerHost + ":" + brokerPort;
@@ -465,6 +474,14 @@ public class MQTTModularInput extends ModularInput {
 		arg.setTitle("Stanza Name");
 		arg.setDescription("");
 
+		endpoint.addArg(arg);
+		
+		arg = new Arg();
+		arg.setName("activation_key");
+		arg.setTitle("Activation Key");
+		arg.setDescription("Visit http://www.baboonbones.com/#activation  to obtain a free,non-expiring key");
+		arg.setRequired_on_create(true);
+		arg.setRequired_on_edit(true);
 		endpoint.addArg(arg);
 
 		arg = new Arg();