+++ protocol/src/com/splunk/modinput/protocol/protocolverticle/HTTPVerticle.java
@@ -32,21 +32,18 @@ public class HTTPVerticle extends Verticle {
 
 		JsonObject config = container.config();
 
-		JsonObject handlerConfig = new JsonObject(
-				config.getString("handler_config"));
+		JsonObject handlerConfig = new JsonObject(config.getString("handler_config"));
 
 		handlerConfig.putString("address", address);
 
-		container.deployWorkerVerticle(config.getString("handler_verticle"),
-				handlerConfig, config.getNumber("handler_verticle_instances")
-						.intValue(), false, new AsyncResultHandler<String>() {
+		container.deployWorkerVerticle(config.getString("handler_verticle"), handlerConfig,
+				config.getNumber("handler_verticle_instances").intValue(), false, new AsyncResultHandler<String>() {
 					public void handle(AsyncResult<String> asyncResult) {
 						if (asyncResult.succeeded()) {
 							// ok
 						} else {
 							logger.error("Can't instantiate handler verticle : "
-									+ ModularInput.getStackTrace(asyncResult
-											.cause()));
+									+ ModularInput.getStackTrace(asyncResult.cause()));
 
 						}
 					}
@@ -55,14 +52,12 @@ public class HTTPVerticle extends Verticle {
 		boolean useSSL = false;
 
 		if (config.containsField("use_ssl"))
-			useSSL = Boolean.parseBoolean(config.getNumber("use_ssl")
-					.intValue() == 1 ? "true" : "false");
+			useSSL = Boolean.parseBoolean(config.getNumber("use_ssl").intValue() == 1 ? "true" : "false");
 
 		HttpServer server = vertx.createHttpServer();
 
 		if (config.containsField("accept_backlog"))
-			server.setAcceptBacklog(config.getNumber("accept_backlog")
-					.intValue());
+			server.setAcceptBacklog(config.getNumber("accept_backlog").intValue());
 
 		if (useSSL) {
 			server.setSSL(true);
@@ -73,23 +68,20 @@ public class HTTPVerticle extends Verticle {
 			if (config.containsField("truststore_path"))
 				server.setTrustStorePath(config.getString("truststore_path"));
 			if (config.containsField("truststore_pass"))
-				server.setTrustStorePassword(config
-						.getString("truststore_pass"));
+				server.setTrustStorePassword(config.getString("truststore_pass"));
 			if (config.containsField("client_auth_required"))
 				server.setClientAuthRequired(Boolean
-						.parseBoolean(config.getNumber("client_auth_required")
-								.intValue() == 1 ? "true" : "false"));
+						.parseBoolean(config.getNumber("client_auth_required").intValue() == 1 ? "true" : "false"));
 		}
 
 		if (config.containsField("receive_buffer_size"))
-			server.setReceiveBufferSize(config.getNumber("receive_buffer_size")
-					.intValue());
+			server.setReceiveBufferSize(config.getNumber("receive_buffer_size").intValue());
 		if (config.containsField("tcp_nodelay"))
-			server.setTCPNoDelay(Boolean.parseBoolean(config.getNumber(
-					"tcp_nodelay").intValue() == 1 ? "true" : "false"));
+			server.setTCPNoDelay(
+					Boolean.parseBoolean(config.getNumber("tcp_nodelay").intValue() == 1 ? "true" : "false"));
 		if (config.containsField("tcp_keepalive"))
-			server.setTCPKeepAlive(Boolean.parseBoolean(config.getNumber(
-					"tcp_keepalive").intValue() == 1 ? "true" : "false"));
+			server.setTCPKeepAlive(
+					Boolean.parseBoolean(config.getNumber("tcp_keepalive").intValue() == 1 ? "true" : "false"));
 		if (config.containsField("so_linger"))
 			server.setSoLinger(config.getNumber("so_linger").intValue());
 
@@ -104,8 +96,7 @@ public class HTTPVerticle extends Verticle {
 
 			public void handle(HttpServerRequest request) {
 				String method = request.method();
-				if (method.equalsIgnoreCase("POST")
-						|| method.equalsIgnoreCase("PUT")) {
+				if (method.equalsIgnoreCase("POST") || method.equalsIgnoreCase("PUT")) {
 
 					request.bodyHandler(new Handler<Buffer>() {
 						public void handle(Buffer body) {
@@ -128,20 +119,32 @@ public class HTTPVerticle extends Verticle {
 						}
 					});
 
-					request.response().setStatusCode(200)
-							.setStatusMessage("OK").end();
+					request.response().setStatusCode(200).setStatusMessage("OK")
+					.putHeader("Access-Control-Allow-Origin", "*")
+					.putHeader("Access-Control-Allow-Methods", "GET,PUT,POST").end();
 
+				}
+
+				else if (method.equalsIgnoreCase("OPTIONS")) {
+					
+					request.bodyHandler(new Handler<Buffer>() {
+						public void handle(Buffer body) {
+							// do nothing
+
+						}
+					});
+					
+					request.response().setStatusCode(200).setStatusMessage("OK")
+							.putHeader("Access-Control-Allow-Origin", "*")
+							.putHeader("Access-Control-Allow-Methods", "GET,PUT,POST").end();
 				} else {
-					request.response()
-							.setStatusCode(405)
-							.setStatusMessage(
-									"Method " + method + " Not Allowed").end();
+
+					request.response().setStatusCode(405).setStatusMessage("Method " + method + " Not Allowed").end();
 					logger.error("Unsupported HTTP Method : " + method);
 				}
 				request.exceptionHandler(new Handler<Throwable>() {
 					public void handle(Throwable t) {
-						logger.error("Error in the HTTP Server: "
-								+ ModularInput.getStackTrace(t));
+						logger.error("Error in the HTTP Server: " + ModularInput.getStackTrace(t));
 					}
 				});
 			}