+++ alexa/docs/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+1.0
+-----
+Added an activation key requirement , visit http://www.baboonbones.com/#activation  to obtain a free,non-expiring key
+Docs updated
+Splunk 7.1 compatible
+
 0.8
 ----
 * Trivial bug fix that was messing with my OCD