+++ jms/appServer/static/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+1.6
+-----
+Added an activation key requirement , visit http://www.baboonbones.com/#activation  to obtain a free,non-expiring key
+Docs updated
+Splunk 7.1 compatible
+
 1.5.1
 -----
 Added a new message handler that just dumps the message body :