+++ jmx/appServer/static/docs/readme/RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+2.5
+-----
+Added an activation key requirement , visit http://www.baboonbones.com/#activation  to obtain a free,non-expiring key
+Docs updated
+Splunk 7.1 compatible
+
 2.4
 ---
 Minor HEC tweaks